Separate master to each protocol.
