+++ rzrtd/src/bgp/mod.rs
@@ -2,7 +2,6 @@
 // ReZe.Rs - Router Daemon
 //   Copyright (C) 2018 Toshiaki Takada
 //
-// OSPF Version 2
-//   Supported specifictions:
-//     RFC2328 TBD
+// Border Gateway Protocol
+//
 pub mod master;
