+++ rzrtd/src/zebra/mod.rs
@@ -4,3 +4,4 @@
 //
 // Zebra - IPv4/IPv6 RIB manager, Kernel Interface.
 //
+pub mod master;
